+++ app/Mail/DemoRequest.php
@@ -12,12 +12,13 @@ class DemoRequest extends Mailable
 {
     use Queueable, SerializesModels;
 
-    public $formData;
+    /** @var array<string, mixed> */
+    public array $formData;
 
     /**
-     * Create a new message instance.
+     * @param  array<string, mixed>  $formData
      */
-    public function __construct($formData)
+    public function __construct(array $formData)
     {
         $this->formData = $formData;
     }
@@ -29,7 +30,7 @@ class DemoRequest extends Mailable
     {
         return new Envelope(
             subject: 'New Demo Request from '.$this->formData['name'],
-            to: 'education@datakind.org'
+            to: ['education@datakind.org']
         );
     }